0.4.1 优化体验 完善文档

+++ src/main/java/io/github/c20c01/cc_mb/item/Awl.java
@@ -33,7 +33,7 @@ public class Awl extends Item {
         CompoundTag tag = awl.getOrCreateTag();
 
         byte tickPerBeat = getTickPerBeatTag(tag);
-        tickPerBeat += player.isShiftKeyDown() ? -1 : 1;
+        tickPerBeat += (byte) (player.isShiftKeyDown() ? -1 : 1);
 
         if (tickPerBeat < MusicBoxBlockEntity.MIN_TICK_PER_BEAT) {
             tickPerBeat = MusicBoxBlockEntity.MAX_TICK_PER_BEAT;
